Created file upload form & upload processing page
upload.php is the page that asks the website user to upload a file. 
getfile.php processes the upload form after the upload button is clicked and saves the selected file in the web root directory. Error checking is in the code, but commented out so the user can currently overwrite whatever is already in that directory. Upload with caution!
header.php has been updated with a link to the upload page.